🔨 refactor: cd player 컴포넌트 분리
